CEM Database Backend Class

An interface to the common evidence model that uses works directly with a database schema

Value

data.frame of sources data.frame of condition concept_id and concept_name data.frame of condition concept_id and concept_name

Super class

CemConnector::AbstractCemBackend -> CemDatabaseBackend

Public fields

connection

ConnectionHandlder instance

cemDatabaseSchema

schema of CEM database

vocabularyDatabaseSchema

OMOP vocabulary schema (must include concept and concept ancestor tables)

sourceDatabaseSchema

schema containing source_info table

Methods

Public methods

CemDatabaseBackend$new()

CemDatabaseBackend$finalize()

CemDatabaseBackend$getConditionEvidenceSummary()

CemDatabaseBackend$getConditionEvidence()

CemDatabaseBackend$getIngredientEvidenceSummary()

CemDatabaseBackend$getIngredientEvidence()

CemDatabaseBackend$getRelationships()

CemDatabaseBackend$getCemSourceInfo()

CemDatabaseBackend$getSuggestedControlCondtions()

CemDatabaseBackend$getSuggestedControlIngredients()

CemDatabaseBackend$clone()

Inherited methods

Method new()

initialize backend object.

Usage

CemDatabaseBackend$new(
  connectionDetails,
  cemDatabaseSchema,
  vocabularyDatabaseSchema,
  sourceDatabaseSchema,
  usePooledConnection = FALSE
)

Arguments

connectionDetails

DatabaseConnector connection details object

cemDatabaseSchema

Schema name including CEM unified and matrix_summary tables

vocabularyDatabaseSchema

OMOP vocabulary

sourceDatabaseSchema

Schema containing CEM source information table

usePooledConnection

Used a pooled connection object rather than a database connector object.

Method finalize()

Closes connection Condition evidence summary

Usage

CemDatabaseBackend$finalize()

Method getConditionEvidenceSummary()

Returns set of ingredient concepts for a given conceptset of outcomes

Usage

CemDatabaseBackend$getConditionEvidenceSummary(
  conditionConceptSet,
  siblingLookupLevels = 0
)

Arguments

conditionConceptSet

data.frame conforming to conceptset format, must be standard SNOMED conditions

siblingLookupLevels

where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump Condition evidence

Method getConditionEvidence()

Returns set of relationships that exist within CEM for a condition conceptset of interest

Usage

CemDatabaseBackend$getConditionEvidence(
  conditionConceptSet,
  siblingLookupLevels = 0
)

Arguments

conditionConceptSet

data.frame conforming to conceptset format, must be standard SNOMED conditions

siblingLookupLevels

where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump Ingredient evidence summary

Method getIngredientEvidenceSummary()

Returns set of outcome concepts for a given conceptset of ingredients/exposures

Usage

CemDatabaseBackend$getIngredientEvidenceSummary(ingredientConceptSet)

Arguments

ingredientConceptSet

data.frame conforming to conceptset format, must be standard RxNorm Ingredients Ingredient conceptset evidence

Method getIngredientEvidence()

for a concept set of rxnorm ingredients (or their ancestors) list the evidence stored in the CEM Utilises matrix summary table for optimisation.

Usage

CemDatabaseBackend$getIngredientEvidence(ingredientConceptSet)

Arguments

ingredientConceptSet

data.frame conforming to conceptset format, must be standard RxNorm Ingredients

Method getRelationships()

From a unified CEM relationships table, for a conceptSet of drug ingredients and a conceptSet of conditions,

Usage

CemDatabaseBackend$getRelationships(
  ingredientConceptSet,
  conditionConceptSet,
  conditionSiblingLookupLevels = 0
)

Arguments

ingredientConceptSet

data.frame conforming to conceptset format, must be standard RxNorm Ingredients

conditionConceptSet

data.frame conforming to conceptset format, must be standard SNOMED conditions

conditionSiblingLookupLevels

integer - where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump

Method getCemSourceInfo()

Get CEM source info as a dataframe

Usage

CemDatabaseBackend$getCemSourceInfo()

Method getSuggestedControlCondtions()

Get negative control snomed condition concepts for a given conceptset These are ranked by co-occurrence accross ohdsi studies A negative control for a submitted concept_set is valid if there is no evidence for the outcome

Usage

CemDatabaseBackend$getSuggestedControlCondtions(
  ingredientConceptSet,
  nControls = 100
)

Arguments

ingredientConceptSet

data.frame conforming to conceptset format, must be standard RxNorm Ingredients

nControls

topN controls to select - the maximum number will be limited by available concepts without related evidence

Method getSuggestedControlIngredients()

Get negative control rxnorm ingredient concepts for a given conceptset These are ranked by co-occurrence accross ohdsi studies A negative control for a submitted concept_set is valid if there is no evidence for the ingredient/condition combination

Usage

CemDatabaseBackend$getSuggestedControlIngredients(
  conditionConceptSet,
  siblingLookupLevels = 0,
  nControls = 100
)

Arguments

conditionConceptSet

data.frame conforming to conceptset format, must be standard SNOMED conditions

siblingLookupLevels

where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump

nControls

topN controls to select - the maximum number will be limited by available concepts without related evidence

Method clone()

The objects of this class are cloneable with this method.

Usage

CemDatabaseBackend$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.