CEM Web Backend Class

An interface to the common evidence model that uses http requests

Value

data.frame of condition concept_id and concept_name data.frame of condition concept_id and concept_name

Super class

CemConnector::AbstractCemBackend -> CemWebApiBackend

Public fields

apiUrl

url for the common evidence model hosted api instance

Methods

Public methods

CemWebApiBackend$new()

CemWebApiBackend$request()

CemWebApiBackend$getStatus()

CemWebApiBackend$getConditionEvidenceSummary()

CemWebApiBackend$getIngredientEvidenceSummary()

CemWebApiBackend$getConditionEvidence()

CemWebApiBackend$getIngredientEvidence()

CemWebApiBackend$getRelationships()

CemWebApiBackend$getCemSourceInfo()

CemWebApiBackend$getVersion()

CemWebApiBackend$getSuggestedControlCondtions()

CemWebApiBackend$getSuggestedControlIngredients()

CemWebApiBackend$clone()

Inherited methods

CemConnector::AbstractCemBackend$finalize()

Method new()

initialize object

Usage

CemWebApiBackend$new(apiUrl)

Arguments

apiUrl

String URL parameter for hosted

Method request()

Do a web request

Usage

CemWebApiBackend$request(method, endpoint, ...)

Arguments

method

string "POST", "GET", "PUT" (not implemented), "DELETE" (not implemented)

endpoint

URL endpoint string

...

list params for httr method

Method getStatus()

GET server status

Usage

CemWebApiBackend$getStatus()

Method getConditionEvidenceSummary()

Returns set of ingredient concepts for a given conceptset of outcomes

Usage

CemWebApiBackend$getConditionEvidenceSummary(
  conditionConceptSet,
  siblingLookupLevels = 0
)

Arguments

conditionConceptSet

data.frame conforming to conceptset format, must be standard SNOMED conditions

siblingLookupLevels

where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump

Method getIngredientEvidenceSummary()

Returns set of outcome concepts for a given conceptset of ingredients

Usage

CemWebApiBackend$getIngredientEvidenceSummary(ingredientConceptSet)

Arguments

ingredientConceptSet

data.frame conforming to conceptset format, must be standard RxNorm Ingredients

Method getConditionEvidence()

Returns ingredient evidence for a given conceptset of conditions

Usage

CemWebApiBackend$getConditionEvidence(
  conditionConceptSet,
  siblingLookupLevels = 0
)

Arguments

conditionConceptSet

data.frame conforming to conceptset format, must be standard SNOMED conditions

siblingLookupLevels

where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump

Method getIngredientEvidence()

Returns condtion evidence for a given conceptset of ingredients

Usage

CemWebApiBackend$getIngredientEvidence(ingredientConceptSet)

Arguments

ingredientConceptSet

data.frame conforming to conceptset format, must be standard RxNorm Ingredients

Method getRelationships()

From a unified CEM relationships table, for a conceptSet of drug ingredients and a conceptSet of conditions,

Usage

CemWebApiBackend$getRelationships(
  ingredientConceptSet,
  conditionConceptSet,
  conditionSiblingLookupLevels = 0
)

Arguments

ingredientConceptSet

data.frame conforming to conceptset format, must be standard RxNorm Ingredients

conditionConceptSet

data.frame conforming to conceptset format, must be standard SNOMED conditions

conditionSiblingLookupLevels

integer - where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump

Method getCemSourceInfo()

Returns data.frame of sources that made the CEM

Usage

CemWebApiBackend$getCemSourceInfo()

Method getVersion()

getVersion information from api

Usage

CemWebApiBackend$getVersion()

Method getSuggestedControlCondtions()

Get negative control snomed condition concepts for a given conceptset These are ranked by co-occurrence accross ohdsi studies A negative control for a submitted concept_set is valid if there is no evidence for the outcome

Usage

CemWebApiBackend$getSuggestedControlCondtions(
  ingredientConceptSet,
  nControls = 100
)

Arguments

ingredientConceptSet

data.frame conforming to conceptset format, must be standard RxNorm Ingredients

nControls

topN controls to select - the maximum number will be limited by available concepts without related evidence

Method getSuggestedControlIngredients()

Get negative control rxnorm ingredient concepts for a given conceptset These are ranked by co-occurrence accross ohdsi studies A negative control for a submitted concept_set is valid if there is no evidence for the ingredient/condition combination

Usage

CemWebApiBackend$getSuggestedControlIngredients(
  conditionConceptSet,
  siblingLookupLevels = 0,
  nControls = 100
)

Arguments

conditionConceptSet

data.frame conforming to conceptset format, must be standard SNOMED conditions

siblingLookupLevels

where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump

nControls

topN controls to select - the maximum number will be limited by available concepts without related evidence

Method clone()

The objects of this class are cloneable with this method.

Usage

CemWebApiBackend$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.